refactor(rate-limit): extract endpoint config table to own module
