+++ .agents/metrics/metrics-opencode.jsonl
@@ -14,3 +14,4 @@
 {"timestamp": "2026-09-05T17:45:00Z", "agent": "opencode", "task": "plans folder cleanup, archive 14 stale docs (PR 757)", "status": "completed"}
 {"timestamp": "2026-09-05T18:10:00Z", "agent": "opencode", "task": "close t-2 t-3 t-4 test gaps, 95 tests (PR 759)", "status": "completed"}
 {"timestamp": "2026-09-07T11:05:00Z", "agent": "opencode", "task": "ci unblock eresolve full hygiene sweep pr780", "status": "completed"}
+{"timestamp": "2026-09-07T11:20:00Z", "agent": "opencode", "task": "metrics v0.19.14 sync pr781", "status": "completed"}
