+++ worker/lib/rate-limit.ts
@@ -13,12 +13,22 @@ import { logger } from "./global-logger";
 import { toErrMessage } from "./errors";
 import { checkRateLimitViaBinding } from "./rate-limit-binding";
 import { listAllKvKeys } from "./kv-pagination";
+import {
+  DEFAULT_CONFIG,
+  ENDPOINT_LIMITS,
+  SENSITIVE_ENDPOINTS,
+  getPerKeyRateLimitConfig,
+} from "./rate-limit-config";
+import type { RateLimitConfig } from "./rate-limit-config";
 
-export interface RateLimitConfig {
-  maxRequests: number;
-  windowSeconds: number;
-  keyPrefix: string;
-}
+export type { RateLimitConfig } from "./rate-limit-config";
+export {
+  DEFAULT_CONFIG,
+  ENDPOINT_LIMITS,
+  SENSITIVE_ENDPOINTS,
+  getPerKeyRateLimitConfig,
+  getRateLimitConfig,
+} from "./rate-limit-config";
 
 export interface RateLimitResult {
   allowed: boolean;
@@ -56,53 +66,9 @@ interface RateLimitState {
   windowStart: number;
 }
 
-const DEFAULT_CONFIG: RateLimitConfig = {
-  maxRequests: 100,
-  windowSeconds: 60,
-  keyPrefix: "ratelimit",
-};
 const DEFAULT_KV_MAX_REQUESTS = 100;
 const DEFAULT_KV_WINDOW_SECONDS = 60;
 const KV_KEY_PREFIX = "rl:kv";
-
-function cfg(
-  maxRequests: number,
-  keySuffix: string,
-  windowSeconds = 60,
-): RateLimitConfig {
-  return { maxRequests, windowSeconds, keyPrefix: "ratelimit:" + keySuffix };
-}
-
-const ENDPOINT_LIMITS: Record<string, RateLimitConfig> = {
-  "/api/submit": cfg(10, "submit"),
-  "/api/discover": cfg(5, "discover", 300),
-  "/api/research": cfg(20, "research"),
-  "/api/email/incoming": cfg(30, "email"),
-  "/api/email/parse": cfg(20, "email-parse"),
-  "/api/validate/url": cfg(20, "validate"),
-  "/api/validate/batch": cfg(5, "validate-batch", 300),
-  "/api/semantic-search": cfg(10, "semantic"),
-  "/api/auth/register": cfg(5, "auth-register"),
-  "/api/auth/login": cfg(10, "auth-login"),
-  "/api/auth/refresh": cfg(20, "auth-refresh"),
-  "/api/nlq": cfg(10, "nlq"),
-  "/api/experience": cfg(20, "experience"),
-  "/deals": cfg(60, "deals"),
-  "/webhooks/incoming": cfg(50, "webhook"),
-  default: DEFAULT_CONFIG,
-};
-
-const SENSITIVE_ENDPOINTS = new Set([
-  "/api/auth/register",
-  "/api/auth/login",
-  "/api/auth/refresh",
-  "/api/submit",
-  "/api/email/incoming",
-  "/api/email/parse",
-  "/api/validate/url",
-  "/api/validate/batch",
-  "/webhooks/incoming",
-]);
 
 /** Check rate limit via binding (primary) or KV (fallback). */
 export async function checkRateLimit(
@@ -461,23 +427,6 @@ export function createRateLimitKVMiddleware(
     response.headers.set("X-RateLimit-Reset", result.resetAt.toISOString());
     return response;
   };
-}
-
-/** Parse per-key rate limit config from AuthResult metadata. */
-export function getPerKeyRateLimitConfig(
-  auth: AuthResult,
-): RateLimitConfig | undefined {
-  if (!auth.requestsPerMinute && !auth.requestsPerHour) return undefined;
-  return {
-    maxRequests: auth.requestsPerMinute ?? DEFAULT_CONFIG.maxRequests,
-    windowSeconds: 60,
-    keyPrefix: "ratelimit:user",
-  };
-}
-
-/** Get rate limit configuration for an endpoint. */
-export function getRateLimitConfig(endpoint: string): RateLimitConfig {
-  return ENDPOINT_LIMITS[endpoint] ?? DEFAULT_CONFIG;
 }
 
 /** Reset rate limit state for a client and endpoint. */
